requirements.txt
python-dotenv
sounddevice
scipy
faster-whisper
openai
gTTS
numpy
